+++ common-lootjs/src/main/java/com/yanny/ali/plugin/lootjs/LootJsPlugin.java
@@ -49,6 +49,7 @@ public class LootJsPlugin implements IPlugin {
                         LOGGER.error("Skipping unexpected modification type {}", lootModification.getClass().getCanonicalName());
                     }
                 } catch (Throwable e) {
+                    e.printStackTrace();
                     LOGGER.error("Failed to process loot modification {}: {}", ((MixinAbstractLootModification) lootModification).getName(), e.getMessage());
                 }
             } else {